feat(compiler): set thread_safe_pointer to false by default and enhance ci (#3690)
BREAKING CHANGE: Rust IDL ref fields now generate Rc/RcWeak unless
thread_safe=true is set, which generates Arc/ArcWeak.



## Why?

Rust IDL ref fields do not need cross-thread ownership by default. Using
`Rc`/`RcWeak` as the default keeps generated Rust types aligned with the
non-thread-safe default while still allowing users to opt into
`Arc`/`ArcWeak` when cross-thread sharing is required.



## What does this PR do?

- Adds the `thread_safe_pointer` field option to the compiler protobuf
extension so protobuf IDL can opt Rust ref fields into `Arc`/`ArcWeak`.
- Updates Rust codegen so ref fields, weak refs, list element refs, map
value refs, and union payload refs default to `Rc`/`RcWeak` unless
`thread_safe=true` is set.
- Updates generated-code tests and Rust IDL roundtrip tests to validate
the new default pointer mapping and the thread-safe opt-in path.
- Refreshes compiler IDL docs for Fory IDL, protobuf IDL, and
FlatBuffers IDL with the new Rust pointer carrier behavior.
- Adds a dedicated Scala IDL CI job, enables warnings-as-errors for the
C++ example builds, and fixes AVX2 constant helpers used by UTF string
utilities.



## Related issues



## AI Contribution Checklist



- [ ] Substantial AI assistance was used in this PR: `yes` / `no`
- [ ] If `yes`, I included a completed [AI Contribution
Checklist](https://github.com/apache/fory/blob/main/AI_POLICY.md#9-contributor-checklist-for-ai-assisted-prs)
in this PR description and the required `AI Usage Disclosure`.
- [ ] If `yes`, my PR description includes the required `ai_review`
summary and screenshot evidence of the final clean AI review results
from both fresh reviewers on the current PR diff or current HEAD after
the latest code changes.



## Does this PR introduce any user-facing change?



- [x] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

+++ cpp/fory/util/string_util.cc
@@ -159,10 +159,14 @@ FORY_TARGET_AVX2_ATTR std::string utf16_to_utf8(const std::u16string &utf16,
 
   const __m256i limit1 = _mm256_set1_epi16(0x80);
   const __m256i limit2 = _mm256_set1_epi16(0x800);
-  const __m256i surrogate_high_start = _mm256_set1_epi16(0xD800);
-  const __m256i surrogate_high_end = _mm256_set1_epi16(0xDBFF);
-  const __m256i surrogate_low_start = _mm256_set1_epi16(0xDC00);
-  const __m256i surrogate_low_end = _mm256_set1_epi16(0xDFFF);
+  const __m256i surrogate_high_start =
+      _mm256_set1_epi16(static_cast<int16_t>(0xD800));
+  const __m256i surrogate_high_end =
+      _mm256_set1_epi16(static_cast<int16_t>(0xDBFF));
+  const __m256i surrogate_low_start =
+      _mm256_set1_epi16(static_cast<int16_t>(0xDC00));
+  const __m256i surrogate_low_end =
+      _mm256_set1_epi16(static_cast<int16_t>(0xDFFF));
 
   char buffer[64]; // Buffer to hold temporary UTF-8 bytes
   char *output = buffer;
